+++ examples/simple_example/kms.tf
@@ -23,7 +23,7 @@ resource "random_string" "key_suffix" {
 resource "google_kms_key_ring" "keyring" {
   project  = var.project_id
   name     = "alloydb-keyring-example-${random_string.key_suffix.result}"
-  location = "us-central1"
+  location = var.region
 }
 
 resource "google_kms_crypto_key" "key" {
@@ -40,6 +40,26 @@ resource "google_project_service_identity" "alloydb_sa" {
 
 resource "google_kms_crypto_key_iam_member" "alloydb_sa_iam" {
   crypto_key_id = google_kms_crypto_key.key.id
+  role          = "roles/cloudkms.cryptoKeyEncrypterDecrypter"
+  member        = "serviceAccount:${google_project_service_identity.alloydb_sa.email}"
+}
+
+
+## Cross Region Secondary Cluster Keys
+
+resource "google_kms_key_ring" "keyring_secondary" {
+  project  = var.project_id
+  name     = "alloydb-keyring-example-sec-${random_string.key_suffix.result}"
+  location = var.secondary_region
+}
+
+resource "google_kms_crypto_key" "key_secondary" {
+  name     = "crypto-key-example-sec-${random_string.key_suffix.result}"
+  key_ring = google_kms_key_ring.keyring_secondary.id
+}
+
+resource "google_kms_crypto_key_iam_member" "alloydb_sa_iam_secondary" {
+  crypto_key_id = google_kms_crypto_key.key_secondary.id
   role          = "roles/cloudkms.cryptoKeyEncrypterDecrypter"
   member        = "serviceAccount:${google_project_service_identity.alloydb_sa.email}"
 }