+++ examples/simple_example/outputs.tf
@@ -33,3 +33,38 @@ output "kms_key_name" {
   description = "he fully-qualified resource name of the KMS key"
   value       = google_kms_crypto_key.key.id
 }
+
+output "cluster_name" {
+  description = "The name of the cluster resource"
+  value       = module.alloy-db.cluster_name
+}
+
+output "region" {
+  description = "The region for primary cluster"
+  value       = var.region
+}
+
+output "secondary_region" {
+  description = "The region for cross region replica secondary cluster"
+  value       = var.secondary_region
+}
+
+output "secondary_cluster_id" {
+  description = "ID of the Secondary Alloy DB Cluster created"
+  value       = module.alloy-db-secondary.cluster_id
+}
+
+output "secondary_primary_instance_id" {
+  description = "ID of the Secondary Cluster primary instance created"
+  value       = module.alloy-db-secondary.primary_instance_id
+}
+
+output "secondary_kms_key_name" {
+  description = "he fully-qualified resource name of the Secondary clusterKMS key"
+  value       = google_kms_crypto_key.key_secondary.id
+}
+
+output "secondary_cluster_name" {
+  description = "The name of the Secondary cluster resource"
+  value       = module.alloy-db-secondary.cluster_name
+}